remove BaseFullEditUserView which was only used by one subclass
